Minor changes have been made to the validity of the session creation process

+++ src/main/java/com/redis_integration_demo/service/SessionService.java
@@ -3,7 +3,9 @@ package com.redis_integration_demo.service;
 import com.redis_integration_demo.dto.CreateSessionRequest;
 import com.redis_integration_demo.dto.SessionResponse;
 import com.redis_integration_demo.model.Session;
+import com.redis_integration_demo.model.User;
 import com.redis_integration_demo.repository.SessionRepository;
+import com.redis_integration_demo.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.stereotype.Service;
@@ -13,10 +15,19 @@ import org.springframework.transaction.annotation.Transactional;
 public class SessionService {
     @Autowired
     private SessionRepository sessionRepository;
+    @Autowired
+    private UserRepository userRepository;
 
     @CachePut(value = "sessions", key = "#result.userId")
     @Transactional
     public SessionResponse createSession(CreateSessionRequest request) {
+        User user = userRepository.findById(request.getUserId())
+                        .orElseThrow(() -> new RuntimeException("Пользователь не найден"));
+
+        if (!user.getPassword().equals(request.getPassword())) {
+            throw new RuntimeException("Неверный пароль");
+        }
+
         sessionRepository.deleteByUserId(request.getUserId());
 
         Session session = new Session();
